initial commit
This script searches thru the .osm xml file and does two things:
1) Finds common street abbreviations: St, Pl, Ct, Blvd and spells them out (i.e. Blvd becomes Boulevard)
2) Finds streets in _Denver only_ that are prefixed with "N " (i.e. "N Galapago St") and removes the "N ". ("N Galapago St" becomes "Galapago Street")

This is a command line script that takes 1 or 2 (2nd is optional) arguments. First argument is the source file, second argument is an optional new output file name. If no second argument is provided then "update" is added to the file root name.